Add in main genie file export
