+++ geniesp/loads.py
@@ -11,33 +11,48 @@ from genie import process_functions
 from geniesp.config import BpcConfig
 
 
-@dataclass
-class Loads():
-    bpc_config: BpcConfig
-    upload: bool = False
-    syn: Synapse
+def get_cbioportal_upload_folders(syn: Synapse, staging_release_folder: str, cohort: str, release: str) -> dict:
+    """Create case lists and release folder"""
+    sp_data_folder = syn.store(
+        Folder(cohort, parentId=staging_release_folder)
+    )
+    release_folder = syn.store(Folder(release, parent=sp_data_folder))
+    # Store in cBioPortal files because there is another folder for
+    # clinical files
+    release_folder = syn.store(
+        Folder("cBioPortal_files", parent=release_folder)
+    )
+    case_lists = syn.store(Folder("case_lists", parent=release_folder))
+    return {"release": release_folder, "case_lists": case_lists}
 
-    @cached_property
-    def cbioportal_folders(self) -> dict:
-        """Create case lists and release folder"""
-        if self.upload:
-            return {}
-        sp_data_folder = self.syn.store(
-            Folder(self.bpc_config.cohort, parentId=self.bpc_config.staging_release_folder)
-        )
-        release_folder = self.syn.store(Folder(self.release, parent=sp_data_folder))
-        # Store in cBioPortal files because there is another folder for
-        # clinical files
-        release_folder = self.syn.store(
-            Folder("cBioPortal_files", parent=release_folder)
-        )
-        case_lists = self.syn.store(Folder("case_lists", parent=release_folder))
-        return {"release": release_folder, "case_lists": case_lists}
 
-    def store(self, parent: str, filepath: str, used_entities: Union[list, None] = None):
-        ent = synapseclient.File(filepath, parent=parent)
-        self.syn.store(
-            ent,
-            executed="https://github.com/Sage-Bionetworks/GENIE-Sponsored-Projects",
-            used=used_entities
-        )
+# @dataclass
+# class Loads():
+#     bpc_config: BpcConfig
+#     upload: bool = False
+#     syn: Synapse
+
+#     @cached_property
+#     def cbioportal_folders(self) -> dict:
+#         """Create case lists and release folder"""
+#         if self.upload:
+#             return {}
+#         sp_data_folder = self.syn.store(
+#             Folder(self.bpc_config.cohort, parentId=self.bpc_config.staging_release_folder)
+#         )
+#         release_folder = self.syn.store(Folder(self.bpc_config.release, parent=sp_data_folder))
+#         # Store in cBioPortal files because there is another folder for
+#         # clinical files
+#         release_folder = self.syn.store(
+#             Folder("cBioPortal_files", parent=release_folder)
+#         )
+#         case_lists = self.syn.store(Folder("case_lists", parent=release_folder))
+#         return {"release": release_folder, "case_lists": case_lists}
+
+#     def store(self, parent: str, filepath: str, used_entities: Union[list, None] = None):
+#         ent = synapseclient.File(filepath, parent=parent)
+#         self.syn.store(
+#             ent,
+#             executed="https://github.com/Sage-Bionetworks/GENIE-Sponsored-Projects",
+#             used=used_entities
+#         )
